+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 3.5.0
-- main() now meaningfully supports external invokation.
+- CLI tool now support a single-source mode, i.e. the setup where there's a single data source
+having all information about all benchmarks across all contenders. This enables CLI tool to produce N-way comparisons whenever needed. Now when the
+`--file1_parser`'s `getAltDelimiter()` method returns a string, the string is used as
+an alternatives delimiter. For the details, see `alt_delimiter=` parameter of
+`benchstats.compare.compareStats()`.
+- main() now meaningfully supports external invocation.
 
 # 3.4.0
 - `qbench.showBench()` can take benchmark results as a dict of per-name 2D `[reps, iters]` arrays